Add files via upload
I took out the print statements for the Eyewall Main and the Eyewall Filter subroutine.  This should increase the speed of the calculations.  I did add a print to the front and back of the Eyewall Main to let the user know when the program has finished compiling.  Lastly, I changed the count in line 35 to just count.  We still run into a problem where it does not index the last file in the list, but we do not run into an index out-of-bounds error anymore.